requirements.txt
# Core
flask==3.1.0
flask-cors==5.0.1

# ML - Semantic Search (recommended)
sentence-transformers==3.4.1
faiss-cpu==1.13.2

# numpy is installed automatically by sentence-transformers
